requirements.txt
pandas 
numpy
matplotlib
seaborn
scikit-learn
nltk
gensim
sentence-transformers
streamlit
python-docx
PyMuPDF